Внесены правки

+++ src/main/java/pro/sky/java/course2/calculator250623/service/CalculatorService.java
@@ -4,5 +4,5 @@ public interface CalculatorService {
     int getPlusResult(int num1, int num2);
     int getMinusResult(int num1, int num2);
     int getMultiplyResult(int num1, int num2);
-    int getDivideResult(int num1, int num2);
+    double getDivideResult(int num1, int num2);
 }
