+++ src/main/java/pro/sky/java/course2/calculator250623/service/CalculatorServiceImpl.java
@@ -18,12 +18,12 @@ public class CalculatorServiceImpl implements CalculatorService {
         return num1 * num2;
     }
     @Override
-    public int getDivideResult(int num1, int num2) {
+    public double getDivideResult(int num1, int num2) {
         if (num2 == 0) {
             throw new IllegalArgumentException("Divide by zero error");
         }
         else {
-            return num1 / num2;
+            return (double) num1 / num2;
         }
     }
 }
